Fixed types to be consistently uintX_t and use r and c

+++ game_of_life/src/worker.h
@@ -4,9 +4,9 @@
 #include "world.h"
 
 // worker thread for calculating the world in strips
-unsafe void worker(bit (*unsafe world)[BITSLOTSP(WDHT + 4, WDWD + 4)], int wnumber, chanend toDist, chanend toNextWorker, chanend fromLastWorker);
+unsafe void worker(bit (*unsafe world)[BITSLOTSP(WDHT + 4, WDWD + 4)], uint8_t wnumber, chanend toDist, chanend toNextWorker, chanend fromLastWorker);
 
 // thread for last worker, doesn't take a chanend toNextWorker
-unsafe void lastWorker(bit (*unsafe world)[BITSLOTSP(WDHT + 4, WDWD + 4)], int wnumber, chanend toDist, chanend fromLastWorker);
+unsafe void lastWorker(bit (*unsafe world)[BITSLOTSP(WDHT + 4, WDWD + 4)], uint8_t wnumber, chanend toDist, chanend fromLastWorker);
 
 #endif
